perf(attn): bf16 paged KV cache (Phase 1)
Store the paged full-attention KV cache in bf16 instead of f32, halving KV
memory. Mirrors vLLM's bf16 flash_attn KV store (flash_attn.py @ e24d1b24).

- runner: full_attn_buf_ is bf16 (uint16_t words); PagedKvCache.dtype = kBF16.
- qwen3_5 FullAttnBlockPaged: down-cast rope'd f32 K and f32 V to bf16 via
  vt::CastBf16 before ReshapeAndCache. Query stays f32; softmax accumulates f32.
- cuda_paged_attn: decouple query dtype (TQ) from KV-cache dtype (TKV) in the
  decode + prefill-flash kernels so an f32 query reads a bf16 cache (Load()
  converts bf16 to f32). Dispatch on (query.dtype, k_cache.dtype).

+++ include/vllm/v1/worker/gpu/runner.h
@@ -207,9 +207,11 @@ class GPUModelRunner final : public ModelRunnerBase {
   std::vector<int32_t> gdn_free_slots_;
 
   // Owned KV-cache backing storage (host at T0) + the views the forward reads.
-  std::vector<std::vector<float>> full_attn_buf_;  // per full-attn layer
-  std::vector<std::vector<float>> ssm_buf_;        // per GDN layer
-  std::vector<std::vector<float>> conv_buf_;       // per GDN layer
+  // full_attn_buf_ is bf16 (uint16_t words) — the paged KV cache stores K/V in
+  // bf16, mirroring vLLM's bf16 flash_attn KV store (halves KV memory).
+  std::vector<std::vector<uint16_t>> full_attn_buf_;  // per full-attn layer (bf16)
+  std::vector<std::vector<float>> ssm_buf_;           // per GDN layer
+  std::vector<std::vector<float>> conv_buf_;          // per GDN layer
   std::vector<PagedKvCache> attn_kv_;
   std::vector<GdnStateCache> gdn_state_;
